pyproject.toml
[tool.poetry]
name = "mirascope-nolita"
version = "0.1.0"
description = "An example of how to use HDR's Nolita framework as a tool in Mirascope"
authors = ["vintro <vince@plasticlabs.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
mirascope = {extras = ["openai"], version = "^0.18.1"}
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
